requirements.txt
asgiref==3.8.1
django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
markdown==3.8
sqlparse==0.5.3
django-allauth==65.9.0
dj-rest-auth